+++ clang-tools-extra/test/clang-tidy/checkers/readability/Inputs/mark-static-var/func.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void func_header();